% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeData.R
\name{makeData}
\alias{makeData}
\title{Simulate data.}
\usage{
makeData(n = 500, errMean = 0, errSD = 1, seed = 1, intercept = 1, xCoef = 3)
}
\arguments{
\item{n}{Selected sample size of simulated data set, a single integer value greater than 0. Default value is 500.}

\item{errMean}{Mean of the normal distribution, from which the error values shall be sampled. Default value is 0.}

\item{errSD}{Standard deviation of the normal distribution, from which the error values shall be sampled. Default value is 1.}

\item{seed}{A single integer value. Setting a seed ensures reproducibility of a once simulated data set.}

\item{intercept}{Intercept of a simulated model output, defaults value is 1.}

\item{xCoef}{Regression coefficient of a simulated model output, defaults to the coefficient 3.}
}
\value{
simData A data.frame with one continuous predictor column x and one continuous outcome column y.
}
\description{
Simulated data set with one continuous predictor and one continuous outcome, select sample size.
}
\examples{
# Simulate data set with sample size of 100
dfSim <- makeData(n=100)
}
\author{
Marcel Miché
}
